Aun me quedan muchas cosas por arreglar, mañana a la tarde ya estará todo organizado

+++ src/BatallaAcapella.java
@@ -8,7 +8,7 @@ public class BatallaAcapella extends Batalla {
 
     @Override
     public float puntuacionSimulaciones(float R) {
-        return (Math.sqrt(R) + 5)/3;
+        return ((float) Math.sqrt(R) + 5)/3;
     }
 
     @Override
